+++ components/share/api/src/main/java/com/flipperdevices/share/api/CryptoStorageApi.kt
@@ -1,6 +1,9 @@
 package com.flipperdevices.share.api
 
+import com.flipperdevices.bridge.dao.api.model.FlipperKey
+import com.flipperdevices.bridge.dao.api.model.FlipperKeyContent
+
 interface CryptoStorageApi {
-    suspend fun upload(data: ByteArray, path: String, name: String): Result<String>
-    suspend fun download(id: String, key: String, name: String): Result<ByteArray>
+    suspend fun upload(flipperKey: FlipperKey): Result<String>
+    suspend fun download(id: String, key: String): Result<FlipperKeyContent>
 }