+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - [REFACTOR] Enable detekt formatting with rules
 - [REFACTOR] Fix gradle warnings (Deprecated API, PredefinedEnumMap)
 - [REFACTOR] Replace some clickable to card function + reuse orange app bar
+- [REFACTOR] Cipher stream for crypto operation
 
 # 1.4.0